% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizerPlus-package.R
\docType{package}
\name{SynthesizerPlus-package}
\alias{SynthesizerPlus}
\alias{SynthesizerPlus-package}
\title{SynthesizerPlus: Synthetic Data Generation with Copulas, Evaluation and Multi-Format I/O}
\description{
Generates synthetic versions of numeric, categorical, mixed-type and time-series data. Marginal distributions are modelled with empirical quantile functions and dependence with Gaussian or Student-t copulas (Sklar, 1959), using the distributional transform (Ruschendorf, 2009, \doi{10.1016/j.jspi.2009.05.030}) so that categorical variables and missing-value patterns take part in the dependence structure. Supports stratified models, multiple synthetic data sets, chunked generation of very large data sets, vector-autoregressive and bootstrap time-series synthesis, augmentation of small training sets for machine learning, and random generation from multivariate distributions and elliptical and Archimedean copulas. Fidelity, utility and disclosure risk are assessed with marginal distances, association matrices, propensity score mean-squared error (Snoke et al., 2018, \doi{10.1111/rssa.12358}), confidence-interval overlap and distance to closest record, with 'ggplot2' visualisations. Data can be read from and written to CSV, TSV, RDS, RData, Parquet, Feather, HDF5, JSON, NDJSON, Excel, SPSS, Stata, SAS and fst files. The empirical-quantile and copula approach builds on ideas of the 'synthesizer' package by Mark van der Loo.
}
\section{Main functions}{

\itemize{
\item Fitting and generation: \code{\link[=fit_synthesizer]{fit_synthesizer()}}, \code{\link[=generate]{generate()}},
\code{\link[=synthesize]{synthesize()}}, \code{\link[=simulate.sp_synthesizer]{simulate.sp_synthesizer()}}, \code{\link[=generate_to_file]{generate_to_file()}}.
\item Disclosure control: \code{\link[=disclosure_risk]{disclosure_risk()}}, \code{\link[=dcr]{dcr()}}, \link{census_sim} data.
\item Closeness to the real data: \code{closeness} in \code{\link[=generate]{generate()}},
\code{\link[=perturbation]{perturbation()}}, \code{\link[=calibrate_closeness]{calibrate_closeness()}}.
\item Time series: \code{\link[=fit_synthesizer.ts]{fit_synthesizer.ts()}}.
\item Machine-learning augmentation: \code{\link[=augment_data]{augment_data()}}, \code{\link[=ts_windows]{ts_windows()}}.
\item Multivariate distributions: \code{\link[=r_mvnorm]{r_mvnorm()}}, \code{\link[=r_mvt]{r_mvt()}}, \code{\link[=r_mvskewnorm]{r_mvskewnorm()}},
\code{\link[=r_dirichlet]{r_dirichlet()}}, \code{\link[=r_mvmixture]{r_mvmixture()}}, \code{\link[=r_copula]{r_copula()}}, \code{\link[=r_mvdist]{r_mvdist()}},
\code{\link[=make_corr]{make_corr()}}.
\item Evaluation: \code{\link[=compare_synthetic]{compare_synthetic()}}, \code{\link[=marginal_metrics]{marginal_metrics()}},
\code{\link[=association_matrix]{association_matrix()}}, \code{\link[=pmse]{pmse()}}, \code{\link[=discriminator_auc]{discriminator_auc()}},
\code{\link[=ci_overlap]{ci_overlap()}}, \code{\link[=dcr]{dcr()}}.
\item Visualisation: \code{\link[=plot_marginals]{plot_marginals()}}, \code{\link[=plot_association]{plot_association()}}, \code{\link[=plot_pairs]{plot_pairs()}},
\code{\link[=plot_qq]{plot_qq()}}, \code{\link[=plot_ts]{plot_ts()}}, \code{\link[=plot_acf]{plot_acf()}}, \code{\link[=plot_dcr]{plot_dcr()}}.
\item Input/output: \code{\link[=read_data]{read_data()}}, \code{\link[=write_data]{write_data()}}, \code{\link[=supported_formats]{supported_formats()}},
\code{\link[=match_types]{match_types()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bogdanoancea/SynthesizerPlus}
  \item Report bugs at \url{https://github.com/bogdanoancea/SynthesizerPlus/issues}
}

}
\author{
\strong{Maintainer}: Bogdan Oancea \email{bogdan.oancea@gmail.com} [copyright holder]

}
\keyword{internal}
